Add javadoc comment, rename config package to securiry and add transaction in the service
