+++ README.md
@@ -24,4 +24,4 @@
 
 ## Security
 1. Create user service to load user from  database and place in package com.nnk.springboot.services
-2. Add configuration class and place in package com.nnk.springboot.config
+2. Add configuration class and place in package com.nnk.springboot.security
