+++ src/main/java/com/nnk/springboot/domain/Users.java
@@ -5,6 +5,13 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
 
 
+/**
+ * Entity representing a user of the application.
+ * <p>
+ *     This class is mapped to the {@code users} table in the database
+ *     and is used for authentication and authorization operations via Spring Security.
+ * </p>
+ */
 @Entity
 @Table(name = "users")
 public class Users {
@@ -23,6 +30,10 @@ public class Users {
     @NotBlank(message = "Role is mandatory")
     private String role;
 
+    /**
+     * Unique user identifier.
+     * @return the user's ID, or {@code null} if the user has not been persisted yet.
+     */
     public Integer getId() {
         return id;
     }
@@ -31,6 +42,10 @@ public class Users {
         this.id = id;
     }
 
+    /**
+     * Username (login) used for authentication.
+     * @return the username
+     */
     public String getUsername() {
         return username;
     }
@@ -39,6 +54,11 @@ public class Users {
         this.username = username;
     }
 
+    /**
+     * Secure user password.
+     * This field is validated and encoded before being stored.
+     * @return password hash, or {@code null} if the user has not been persisted yet.
+     */
     public String getPassword() {
         return password;
     }
@@ -47,6 +67,10 @@ public class Users {
         this.password = password;
     }
 
+    /**
+     * User's full name (for display purposes).
+     * @return full name, or {@code null} if the user has not been persisted yet.
+     */
     public String getFullname() {
         return fullname;
     }
@@ -55,6 +79,11 @@ public class Users {
         this.fullname = fullname;
     }
 
+    /**
+     * User role in the application (eg: ADMIN, USER).
+     * Used to determine access rights via Spring Security.
+     * @return the role of the user
+     */
     public String getRole() {
         return role;
     }
