+++ src/main/java/com/nnk/validator/PasswordValidator.java
@@ -3,8 +3,28 @@ package com.nnk.validator;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
+/**
+ * Custom validator for user passwords.
+ * <p>
+ *     Ce validateur est déclenché par l'annotation {@link ValidPassword}
+ *     and allows you to check that the password complies with a security policy:
+ * </p>
+ * <ul>
+ *    <li>Not empty</li>
+ *    <li>Minimum 8 characters</li>
+ *   <li>At least one uppercase letter</li>
+ *   <li>At least one special character (!@#$%^&*)</li>
+ * </ul>
+ */
 public class PasswordValidator implements ConstraintValidator<ValidPassword, String> {
 
+    /**
+     * Checks if the password is valid according to the defined security rules.
+     *
+     * @param password the password to validate
+     * @param context the validation context to construct a custom error message
+     * @return {@code true} if the password is correct, {@code false} otherwise
+     */
     @Override
     public boolean isValid(String password, ConstraintValidatorContext context) {
 
@@ -31,6 +51,12 @@ public class PasswordValidator implements ConstraintValidator<ValidPassword, Str
         return true;
     }
 
+    /**
+     * Constructs a custom violation message in the validation context.
+     *
+     * @param context the validation context
+     * @param message the error message to display
+     */
     private void buildConstraint(ConstraintValidatorContext context, String message) {
         context.disableDefaultConstraintViolation();
         context.buildConstraintViolationWithTemplate(message)
